Refactor the classes into seperated classes
