+++ src/main/java/Task/Deadline.java
@@ -1,4 +1,4 @@
-package Deadline;
+package Task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
